Create Functions for variables
A new data set includes many informations that has various columns and rows and it tells a story that what we need to understand. This understanding process can be seperated four parts that are general impression, classification of variables, analysis of categorical variables, analysis of numerical variables, analysis of target variable, analysis of correlation.